Refactored renderer and mesher

+++ src/mesher/placeSection.cpp
@@ -1,22 +1,22 @@
-#include "Mesher.hpp"
+#include "mesher/Mesher.hpp"
 
-std::vector<glm::vec3> Mesher::placeSection(std::vector<glm::vec3> section, glm::vec3& point, glm::vec3& vertexNormal) {
-    glm::mat4 rotationMatrix;
-    glm::mat4 translationMatrix = glm::translate(glm::mat4(1.0f), point);
-    glm::vec3 zAxis = glm::vec3(0.0f, 0.0f, 1.0f);
-    glm::vec3 yAxis = glm::vec3(0.0f, 1.0f, 0.0f);
+std::vector<glm::dvec3> Mesher::placeSection(std::vector<glm::dvec3> section, glm::dvec3& point, glm::dvec3& vertexNormal) {
+    glm::dmat4 rotationMatrix;
+    glm::dmat4 translationMatrix = glm::translate(glm::dmat4(1.0), point);
+    glm::dvec3 zAxis = glm::dvec3(0.0, 0.0, 1.0);
+    glm::dvec3 yAxis = glm::dvec3(0.0, 1.0, 0.0);
 
-    float angle = std::atan2(glm::dot(glm::cross(yAxis, vertexNormal), zAxis), glm::dot(yAxis, vertexNormal));
+    double angle = std::atan2(glm::dot(glm::cross(yAxis, vertexNormal), zAxis), glm::dot(yAxis, vertexNormal));
     if (abs(angle) < 1e-3) {
-        rotationMatrix = glm::mat4(1.0f);
+        rotationMatrix = glm::dmat4(1.0);
     } else {
-        rotationMatrix = glm::rotate(glm::mat4(1.0f), angle, zAxis);
+        rotationMatrix = glm::rotate(glm::dmat4(1.0), angle, zAxis);
     }
 
     glm::mat4 transformationMatrix = translationMatrix * rotationMatrix;
 
-    for (glm::vec3 &p : section) {
-        p = transformationMatrix * glm::vec4(p, 1.0f);
+    for (glm::dvec3 &p : section) {
+        p = transformationMatrix * glm::dvec4(p, 1.0);
     }
     
     return section;
