+++ src/mesher/utils.hpp
@@ -1,11 +1,10 @@
 #pragma once
-
-#include <iostream>
-#include <chrono>
-#include <vector>
 #include <algorithm>
+#include <chrono>
 #include <glm/glm.hpp>
+#include <iostream>
+#include <vector>
 
-float lerp(float start, float end, float alpha);
-int getThroatIndex(const std::vector<glm::vec3>& contour);
-glm::vec3 intersectRaySphere(glm::vec3 rayOrigin, glm::vec3 rayDirection, glm::vec3 sphereOrigin, float sphereRadius);
+double lerp(double start, double end, double alpha);
+int getThroatIndex(const std::vector<glm::dvec3>& contour);
+glm::dvec3 intersectRaySphere(glm::dvec3 rayOrigin, glm::dvec3 rayDirection, glm::dvec3 sphereOrigin, double sphereRadius);
